+++ core/src/main/java/dev/lucianosantos/intervaltimer/core/utils/ICountDownTimerHelper.kt
@@ -1,7 +1,7 @@
 package dev.lucianosantos.intervaltimer.core.utils
 
 interface ICountDownTimerHelper {
-    fun startCountDown(seconds: Long, onTickCallback: (secondsUntilFinished: Long) -> Unit, onFinishCallback: () -> Unit)
+    fun startCountDown(seconds: Int, onTickCallback: (secondsUntilFinished: Long) -> Unit, onFinishCallback: () -> Unit)
 
     fun pause()
